requirements.txt
pyyaml>=6.0
yq>=2.15.0
flake8
black
fabric>=2.6.0
